Separated expenses and incomes frontend
