+++ src/main/java/com/faf223/expensetrackerfaf/web/src/routes/dashboard/board/ExpenseDashboard.svelte
@@ -1,0 +1,9 @@
+<script>
+    import DashHeader from "./expenses/other/DashHeader.svelte";
+    import DataMenu from "./expenses/other/DataMenu.svelte";
+    import QuickInfobar from "./expenses/other/QuickInfobar.svelte";
+</script>
+
+<DashHeader />
+<QuickInfobar />
+<DataMenu />
