+++ src/main/java/com/faf223/expensetrackerfaf/web/src/routes/dashboard/board/IncomeDashboard.svelte
@@ -1,0 +1,9 @@
+<script>
+    import DashHeader from "./incomes/other/DashHeader.svelte";
+    import DataMenu from "./incomes/other/DataMenu.svelte";
+    import QuickInfobar from "./incomes/other/QuickInfobar.svelte";
+</script>
+
+<DashHeader />
+<QuickInfobar />
+<DataMenu />
