+++ src/main/java/com/faf223/expensetrackerfaf/web/src/routes/dashboard/board/expenses/graphs/Graph3.svelte
@@ -1,7 +1,7 @@
 <script>
 	import Chart from 'chart.js/auto';
 	import { onMount } from 'svelte';
-	import { incomeData, expenseData } from "../../stores.js";
+	import { incomeData, expenseData } from "../../../stores.js";
 
 	let ctx;
 	let chartCanvas;
