+++ src/main/java/com/faf223/expensetrackerfaf/web/src/routes/dashboard/stores.js
@@ -7,3 +7,5 @@ export const expenseData = writable([]);
 export const incomeTypes = writable([]);
 
 export const expenseTypes = writable([]);
+
+export let selectedTab = writable('expenses');
